pyproject.toml
[project]
name = "gcal-mcp"
version = "0.1.0"
description = "Google Calendar Events MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.6",
    "google-api-python-client>=2.177.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
